Evaluaciones/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
motor==3.4.0
pymongo==4.6.2  
openai==1.25.1
python-dotenv==1.0.1
